Refactored `ReturnedBasket`
Moved `ReturnedBasket` from being a nested class of `QueryServicer` to its own class to fix the bug that one cannot instantiate a nested class if in a static context (which the `returnBasketId` method is).
