+++ SSH-Cloud/src/main/java/ssh/ReturnedBasket.java
@@ -1,0 +1,31 @@
+package ssh;
+
+import ssh.entities.BasketItem;
+
+import java.util.List;
+
+public class ReturnedBasket{
+    // attributes
+    private Integer basketid;
+    private List<BasketItem> basketItems;
+
+    // constructor
+    public ReturnedBasket(Integer basketid, List<BasketItem> basketItems){
+        this.basketid = basketid;
+        this.basketItems = basketItems;
+    }
+
+    // methods
+    public Integer getBasketid() {
+        return basketid;
+    }
+    public void setBasketid(Integer basketid) {
+        this.basketid = basketid;
+    }
+    public List<BasketItem> getBasketItems() {
+        return basketItems;
+    }
+    public void setBasketItems(List<BasketItem> basketItems) {
+        this.basketItems = basketItems;
+    }
+}
